Support tool call details for clojure-mcp
This changes the `get-tool-call-details` into a multi method that can be
specialized by tool name. It also adds a `get-tool-call-details-result`.

The `get-tool-call-details` method is called BEFORE the tool is called
and the tool gets the chance to add information to inform the user
that the tool is about to be called.

The `get-tool-call-details-result` method is called AFTER the tool is
called and the tool gets the chance to add aditional information to
the details from the `get-tool-call-details` and the actual result.

I tested this with the following prompts, and I do now see pretty
diffs in the chat buffer.

```
create the test.clj file with the content "{:version "1.0.0"}"
bump the version in test.clj. you must use the file_edit MCP tool
bump the version in test.clj. you must use the file_write MCP tool
bump the version in test.clj. you must use the clojure_edit_replace_sexp MCP tool
bump the version in test.clj. you must use the clojure_edit MCP too
```

This PR needs the following change in Emacs (and possibly other clients):

https://github.com/editor-code-assistant/eca-emacs/pull/19