Adjusted to upstream changes.
The upstream change that mostly impacted this commit is that the
config was changed to allow a defaull approval of either :deny,
:allow, or :ask.  This impacted the way manual approvals and
auto-approvals were handled.

This necessitated the addition of a new action
:deliver-default-approval in the state machine table.

During the rebase, I didn't do a good job of properly eliminating the
conflicts. So, a part of this commit is fixing the things I should
have in the rebase.

Removed an unneeded :manual-approve from the event-data of all
:tool-prepare transitions.

Added a :default-approval to the event-data of :auto-approve
transitions.