Fixed invalid state transition bug.
This bug was caused by an incomplete merging of changes to master to
allow the config to auto-allow or auto-deny tool calls.  There was a
state transition from :check-approval to :execution-approved on
:auto-approve.  However, since :auto-approve can now be either allow
or deny, the next-state is not correct.  Since the next state depends
on a conditional, that has tobe in the code and can only influence the
outcome of the state machine via events.  So made the code decide to
send different events depending on the config values.

Updated the tests to accommodate.