Handle transition in :waiting-approval with a :stop-requested event.
This is a work-in-progress checkin.  Not everything works just yet.
In particular, there are concerns that the approval promise will not
be derefed if we get a :stop-requested and will cause a memory leak.
Also, tests have not been updated yet.  I feel that the current tests
are too brittle and they need to be changed or removed.

Removed the unused :deliver-default-approval action.

Changed the name of the :record-metrics action to :log-metrics for
consistency and accuracy.

Updated the state machine to go to :rejected, rather than :stopped, on
a :stop-requested event.

Added guards to proceed down the execution path only if not in
:stopped.

Added actions to store the reason for the tool call decision - both
for allows and denys.