Send as much information as possible with toolCallRejected.
In order to do that, we have to store the necessary information in the
tool call state, as early as possible.  In general, however, the
arguments to the tool call are not available until the :tool-run
event, so the recepients of toolCallRejected have to be able to handle
a null value of the arguments.

Replaced the :init-decision-reason action by the :init-tool-call-state
actions which initializes other aspects of the tool call as well.

Added a :init-arguments action to initialize the arguments to the tool
call when they are known during the :tool-run event.

Added comments specifying that all :send-* actions should typically be
last, and that the status is updated before any actions are run.