Moved the future in :on-tools-called and stored it in the db.
The future in the :on-tools-called callback used to be just before the
check for @approved*.  There is really no need to have a future when
the call has been rejrected, only when the call has been allowed and
ready to go.  So, the future was moved to just before the call.

To prepare for being able to cancel the call future, the future needs
to be stored in the db.  This is done by a new action
`set-call-future` that is called on the :execution-start event.
However, the since the future starts running as soon as it s created,
we have to first delay the future and then force it only after the
tool call state has been transitioned to :executing.