Add :stopping status. Pass state fns to tool calls.
This is an interim commit to eventually allwo tool calls to be
interrupted.

In preparation, we added a :stopping status in between :executing and
:stopped.  This allows us to have self transitions to create and
destroy tool call resources.  The transitions and events are there,
but currently they are not being exercised.  The :stopped status is
considered an "inactive" status similar to an initial or terminal
status.

There is one semantic behavioral change.  Previously, when a tool call
was :executing and it got a :stop-requested event, that was an invalid
transition.  Now, the transition happens and after the tool call is
completed, a toolCallReject message is sent to the client.  Note that
ECA will still wait for the tool call to complete and will not
interrupt it - yet.  That is for a subsequent change.

To allow a tool call to update the tool call state, we have added two
functions that are passed along to the tools.  One function retrieves
the current state of the tool call and the other updates it.
Currently, these functions are passed along to lower level functions
but are not used.

Added tests for the new status and transitions.